Added laravel Gate

+++ app/Http/Controllers/ProductController.php
@@ -7,11 +7,14 @@ use App\Product;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\ProductCreateRequest;
+use Gate;
 
 class ProductController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view', 'products');
+
         $products = Product::paginate();
 
         return ProductResource::collection($products);
@@ -19,11 +22,14 @@ class ProductController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view', 'products');
+
         return new ProductResource((Product::find($id)));
     }
 
     public function store(ProductCreateRequest $request)
     {
+        Gate::authorize('edit', 'products');
 
         $product = Product::create($request->only('title', 'description', 'image', 'price'));
 
@@ -32,6 +38,8 @@ class ProductController extends Controller
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::find($id);
 
         $product->update($request->only('title', 'description', 'image', 'price'));
@@ -41,6 +49,8 @@ class ProductController extends Controller
 
     public function delete($id)
     {
+        Gate::authorize('edit', 'products');
+
         Product::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
