+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\RoleResource;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -16,6 +17,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', 'roles');
+
         return RoleResource::collection(Role::all());
     }
 
@@ -27,6 +30,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
         if ($permissions = $request->input('permissions')) {
@@ -49,6 +54,8 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
+
         return new RoleResource(Role::find('id'));
     }
 
@@ -61,6 +68,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
@@ -87,6 +96,8 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
+
         Role::destroy($id);
 
         \DB::table('role_permission')->where('role_id', $id)->delete();
